// src/app/model/studentModel.ts
export class studentModel {
    student_prefix: any;
    student_firstname: any;
    student_surname: any;
    student_grade: any;
    student_id_card: any;
    student_nickname: any;
    student_residence: any;
    student_room: any;
    student_road: any;
    student_moo: any;
    student_tambon: any;
    student_district: any;
    student_province: any;
    student_province_post: any;
    generate: any;
}